+++ src/solidity-basic/coin.sol
@@ -4,26 +4,24 @@ pragma solidity 0.8.25;
 
 contract Coin {
     address public minter;
-    mapping(address => uint) balances;
+    mapping(address => uint256) balances;
 
-    event Sent(address from, address to, uint amount);
-    error InsufficientBalance(uint required, uint available);
+    event Sent(address from, address to, uint256 amount);
+
+    error InsufficientBalance(uint256 required, uint256 available);
 
     constructor() {
         minter = msg.sender;
     }
 
-    function mint(address receiver, uint amount) public {
+    function mint(address receiver, uint256 amount) public {
         require(msg.sender == minter);
         balances[receiver] += amount;
     }
 
-    function send(address receiver, uint amount) public {
+    function send(address receiver, uint256 amount) public {
         if (balances[msg.sender] < amount) {
-            revert InsufficientBalance({
-                required: amount,
-                available: balances[msg.sender]
-            });
+            revert InsufficientBalance({required: amount, available: balances[msg.sender]});
         }
 
         balances[msg.sender] -= amount;
@@ -32,7 +30,7 @@ contract Coin {
         emit Sent(msg.sender, receiver, amount);
     }
 
-    function getBalance() public view returns (uint){
+    function getBalance() public view returns (uint256) {
         return balances[msg.sender];
     }
 }
